eric has n=7 tasks to complete. his manager gives him m=6 notes on the order tasks must be performed. here is a graph of the dependencies. the dependent array, a = [1,2,3,4,6,5]. his principal tasks array, b = [7,6,4,1,2,1]
it is easy to see that tsk 6 must be performed before task 2 and vice versa. he cna only complete one of those two tasks before the other, so he mush coose either task 6 or 2. he can complete 6 tasks.

the first line contains an int n, the number of tasks.
the next line contains an int m, the size of the array a
each of the next m lines contains an int a[i] where 0 <= i < n
the next line contains an int m, the size of the array a
each of the next m lines contains an int b[i] where 0 <= i < n